change size_t to uint64_t .inl / .h used by upmem device code. read and write in the affine accessor have been changed to support the new blocking notion. also start of dma reduction.

+++ runtime/realm/upmem/legion_c_upmem.h
@@ -24,8 +24,8 @@
 #define LEGION_MAX_DIM 3
 
 typedef long long coord_t;
-typedef size_t FieldID;
-typedef size_t DomainPoint;
+typedef uint64_t FieldID;
+typedef uint64_t DomainPoint;
 
 typedef enum legion_privilege_mode_t
 {
@@ -58,9 +58,9 @@ namespace Legion {
    *  - FT read(const Point<N,T>&) const
    *  ------ Methods below here for [Multi-]Affine Accessors only ------
    *  - const FT* ptr(const Point<N,T>&) const
-   *  - const FT* ptr(const Rect<N,T>&, size_t = sizeof(FT)) const (dense)
-   *  - const FT* ptr(const Rect<N,T>&, size_t strides[N],
-   *                  size_t=sizeof(FT)) const
+   *  - const FT* ptr(const Rect<N,T>&, uint64_t = sizeof(FT)) const (dense)
+   *  - const FT* ptr(const Rect<N,T>&, uint64_t strides[N],
+   *                  uint64_t=sizeof(FT)) const
    *  - const FT& operator[](const Point<N,T>&) const
    *
    * READ_WRITE
@@ -68,8 +68,8 @@ namespace Legion {
    *  - void write(const Point<N,T>&, FT val) const
    *  ------ Methods below here for [Multi-]Affine Accessors only ------
    *  - FT* ptr(const Point<N,T>&) const
-   *  - FT* ptr(const Rect<N,T>&, size_t = sizeof(FT)) const (must be dense)
-   *  - FT* ptr(const Rect<N,T>&, size_t strides[N], size_t=sizeof(FT)) const
+   *  - FT* ptr(const Rect<N,T>&, uint64_t = sizeof(FT)) const (must be dense)
+   *  - FT* ptr(const Rect<N,T>&, uint64_t strides[N], uint64_t=sizeof(FT)) const
    *  - FT& operator[](const Point<N,T>&) const
    *  - template<typename REDOP, bool EXCLUSIVE>
    *      void reduce(const Point<N,T>&, REDOP::RHS) const
@@ -78,8 +78,8 @@ namespace Legion {
    *  - void write(const Point<N,T>&, FT val) const
    *  ------ Methods below here for [Multi-]Affine Accessors only ------
    *  - FT* ptr(const Point<N,T>&) const
-   *  - FT* ptr(const Rect<N,T>&, size_t = sizeof(FT)) const (must be dense)
-   *  - FT* ptr(const Rect<N,T>&, size_t strides[N], size_t=sizeof(FT)) const
+   *  - FT* ptr(const Rect<N,T>&, uint64_t = sizeof(FT)) const (must be dense)
+   *  - FT* ptr(const Rect<N,T>&, uint64_t strides[N], uint64_t=sizeof(FT)) const
    *  - FT& operator[](const Point<N,T>&) const
    */
   template <PrivilegeMode MODE, typename FT, int N, typename COORD_T = coord_t,
